Espansione della funzionalità della gestione fusi orari (offset UTC double)
Dropdown list dei nomi dei paesi in Iglese, oppure nella lingua del paese

+++ SharedData/SqLite/Sqlite_SqlToCreateDatabase.cs
@@ -162,7 +162,7 @@ CREATE TABLE FoodsInMeals (IdFoodInMeal INT NOT NULL, IdMeal INT, IdFood INT, Na
 
 -- Table: GlucoseRecords
 DROP TABLE IF EXISTS GlucoseRecords;
-CREATE TABLE GlucoseRecords (IdGlucoseRecord INT NOT NULL, GlucoseValue DOUBLE, TimeOfMeasurement DATETIME, GlucoseString VARCHAR (45), IdTypeOfGlucoseMeasurement INT, IdOfDevice INT, IdTypeOfDevice INT, IdDeviceModel INT, Notes VARCHAR (255), UtcOffset INTEGER, PRIMARY KEY (IdGlucoseRecord));
+CREATE TABLE GlucoseRecords (IdGlucoseRecord INT NOT NULL, GlucoseValue DOUBLE, TimeOfMeasurement DATETIME, GlucoseString VARCHAR (45), IdTypeOfGlucoseMeasurement INT, IdOfDevice INT, IdTypeOfDevice INT, IdDeviceModel INT, Notes VARCHAR (255), UtcOffset REAL, PRIMARY KEY (IdGlucoseRecord));
 
 -- Table: HypoPredictions
 DROP TABLE IF EXISTS HypoPredictions;
@@ -248,7 +248,7 @@ CREATE TABLE Ingredients (
 
 -- Table: Injections
 DROP TABLE IF EXISTS Injections;
-CREATE TABLE Injections (IdInjection INT NOT NULL, Timestamp DATETIME, InsulinValue DOUBLE, InsulinCalculated DOUBLE, InjectionPositionX DOUBLE, InjectionPositionY DOUBLE, Notes VARCHAR (255), IdTypeOfInjection INT, IdTypeOfInsulinAction INT, IdInsulinDrug INT, InsulinString VARCHAR (45), Zone INTEGER, UtcOffset INTEGER, PRIMARY KEY (IdInjection));
+CREATE TABLE Injections (IdInjection INT NOT NULL, Timestamp DATETIME, InsulinValue DOUBLE, InsulinCalculated DOUBLE, InjectionPositionX DOUBLE, InjectionPositionY DOUBLE, Notes VARCHAR (255), IdTypeOfInjection INT, IdTypeOfInsulinAction INT, IdInsulinDrug INT, InsulinString VARCHAR (45), Zone INTEGER, UtcOffset REAL, PRIMARY KEY (IdInjection));
 
 -- Table: InsulinDrugs
 DROP TABLE IF EXISTS InsulinDrugs;
@@ -277,7 +277,7 @@ CREATE TABLE Manufacturers (
 
 -- Table: Meals
 DROP TABLE IF EXISTS Meals;
-CREATE TABLE Meals (IdMeal INT NOT NULL, IdTypeOfMeal INT, Carbohydrates DOUBLE, TimeBegin DATETIME, Notes VARCHAR (255), AccuracyOfChoEstimate DOUBLE, IdBolusCalculation INT, IdGlucoseRecord INT, IdInjection INT, TimeEnd DATETIME, UtcOffset INTEGER, PRIMARY KEY (IdMeal));
+CREATE TABLE Meals (IdMeal INT NOT NULL, IdTypeOfMeal INT, Carbohydrates DOUBLE, TimeBegin DATETIME, Notes VARCHAR (255), AccuracyOfChoEstimate DOUBLE, IdBolusCalculation INT, IdGlucoseRecord INT, IdInjection INT, TimeEnd DATETIME, UtcOffset REAL, PRIMARY KEY (IdMeal));
 
 -- Table: Parameters
 DROP TABLE IF EXISTS Parameters;
@@ -428,7 +428,7 @@ CREATE TABLE Parameters (
     Time_IsDaylightSavingTime                        INTEGER,
     Time_CountryName                                 VARCHAR(64),
 
-    UtcOffset                                        INTEGER,
+    UtcOffset                                        REAL,
 
     PRIMARY KEY (
